# src/swaggerOutput.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Projeto Trainee Backend - API",
    "description": "Projeto para capacitação trainee de backend, com o objetivo de criar uma API para gerenciar projetos e tarefas."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearer_token": {
      "type": "http",
      "scheme": "bearer",
      "bearerFormat": "JWT",
      "name": "bearer_token",
      "in": "header"
    }
  },
  "consumes": [
    "application/json",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/user/register": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint para cadastrar um usuário",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário criado com sucesso."
          },
          "400": {
            "description": "Erro ao registrar usuário."
          },
          "500": {
            "description": "Erro ao consutar banco de dados."
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint para logar um usuário",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Senha inválida."
          },
          "404": {
            "description": "Usuário não encontrado."
          },
          "500": {
            "description": "Erro ao logar."
          }
        }
      }
    },
    "/user/forgot-password": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint para recuperar a senha de um usuário",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Erro ao enviar email de recuperação de senha."
          },
          "404": {
            "description": "Erro ao encontrar usuário."
          },
          "500": {
            "description": "Erro no esqueci minha senha."
          }
        }
      }
    },
    "/user/reset-password": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint para resetar a senha de um usuário",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "token": {
                  "example": "any"
                },
                "newPassword": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Senha alterada com sucesso."
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Usuário não encontrado."
          },
          "500": {
            "description": "Erro ao resetar senha."
          }
        }
      }
    },
    "/user/userImage/{userId}": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint para adicionar uma imagem a um usuário",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do usuário."
          }
        ],
        "responses": {
          "200": {
            "description": "Imagem registrada com sucesso."
          },
          "400": {
            "description": "Erro ao registrar imagem do usuário."
          },
          "500": {
            "description": "Erro ao registrar imagem do usuário."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/user/turn-moderator": {
      "put": {
        "tags": [
          "User"
        ],
        "description": "Endpoint para tornar um usuário moderador",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário moderador alterado com sucesso."
          },
          "404": {
            "description": "Usuário não encontrado."
          },
          "500": {
            "description": "Erro ao tornar usuário moderador."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/user/delete/{userId}": {
      "delete": {
        "tags": [
          "User"
        ],
        "description": "Endpoint para deletar um usuário",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do usuário."
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário removido com sucesso."
          },
          "400": {
            "description": "Erro ao remover usuário."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    }
  },
...
}